Separate cli from library (#14)
* Move cli to separate file

* Add index.js with exports of required functions

* Cleanup exports

* Update dependencies
