+++ src/index.js
@@ -1,16 +1,8 @@
-#! /usr/bin/env node
-
-const program = require('commander')
-const startMockServer = require('./server')
+const { createMockServer, startMockServer } = require('./server')
 const mockSchema = require('./mock')
 
-program
-    .option('-f --file <file>', 'schema file location', 'schema.graphql')
-    .option('-p --port <port>', 'apollo server port', '4000')
-
-program.parse(process.argv)
-
-const options = program.opts()
-
-const schema = mockSchema(options.file)
-startMockServer(schema, options.port)
+module.exports = {
+    mockSchema,
+    createMockServer,
+    startMockServer
+}
